requirements.txt
# Agent OS Dependencies
# =====================
# Install with: pip install -r requirements.txt

# Core dependencies (required)
python-dotenv==1.0.0    # Load environment variables from .env file
requests==2.31.0        # HTTP client for API calls

# Development dependencies (optional, for testing)
# Install with: pip install -r requirements.txt pytest
# pytest==7.4.0         # Uncomment to install test framework
